requirements.txt
flask==2.3.3
flask-cors==4.0.0
osmnx==1.9.1
geopandas==0.14.3
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
shapely==2.0.3
svgutils==0.3.4
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
flask-jwt-extended==4.6.0
bcrypt==4.1.2
flask-migrate==4.0.5
python-dotenv==1.0.1